feat(refactor): providerコンポーネントとして切り出し
